Update numberQuiz.java
look at the updated asl database file and create the new table that i added for this class. i'm working on fixing a logical error with the questions. when you submit a correct answer, the program will always display a message saying an answer is correct even though it may not be and the same thing happens if it is an incorrect answer being submitted.